Split rbm_spin_real into header and source files
